--- a/examples/armv4t/emu.rs
+++ b/examples/armv4t/emu.rs
@@ -2,6 +2,7 @@ use armv4t_emu::{reg, Cpu, ExampleMem, Memory, Mode};
 
 use crate::mem_sniffer::{AccessKind, MemSniffer};
 use crate::DynResult;
+use gdbstub::common::Pid;
 
 const HLE_RETURN_ADDR: u32 = 0x12345678;
 
@@ -35,6 +36,8 @@ pub struct Emu {
     pub(crate) watchpoints: Vec<u32>,
     pub(crate) breakpoints: Vec<u32>,
     pub(crate) files: Vec<Option<std::fs::File>>,
+
+    pub(crate) reported_pid: Pid,
 }
 
 impl Emu {
@@ -85,6 +88,8 @@ impl Emu {
             watchpoints: Vec::new(),
             breakpoints: Vec::new(),
             files: Vec::new(),
+
+            reported_pid: Pid::new(1).unwrap(),
         })
     }