--- a/src/stub/core_impl/base.rs
+++ b/src/stub/core_impl/base.rs
@@ -34,17 +34,26 @@ impl<T: Target, C: Connection> GdbStubImpl<T, C> {
         Ok(tid)
     }
 
-    pub(crate) fn get_sane_current_pid(
+    pub(crate) fn get_current_pid(
         &mut self,
         target: &mut T,
     ) -> Result<Pid, Error<T::Error, C::Error>> {
-        match target.base_ops() {
-            BaseOps::SingleThread(_) => Ok(FAKE_PID),
-            BaseOps::MultiThread(ops) => ops.active_pid().map_err(Error::TargetError),
+        if let Some(ops) = target
+            .support_extended_mode()
+            .and_then(|ops| ops.support_current_active_pid())
+        {
+            ops.current_active_pid().map_err(Error::TargetError)
+        } else {
+            Ok(FAKE_PID)
         }
     }
 
-    #[inline(always)]
+    // Used by `?` and `vAttach` to return a "reasonable" stop reason.
+    //
+    // This is a bit of an implementation wart, since this is really something
+    // the user ought to be able to customize.
+    //
+    // Works fine for now though...
     pub(crate) fn report_reasonable_stop_reason(
         &mut self,
         res: &mut ResponseWriter<'_, C>,
@@ -58,7 +67,7 @@ impl<T: Target, C: Connection> GdbStubImpl<T, C> {
                 pid: self
                     .features
                     .multiprocess()
-                    .then_some(SpecificIdKind::WithId(self.get_sane_current_pid(target)?)),
+                    .then_some(SpecificIdKind::WithId(self.get_current_pid(target)?)),
                 tid: SpecificIdKind::WithId(tid),
             })?;
         } else {
@@ -190,9 +199,11 @@ impl<T: Target, C: Connection> GdbStubImpl<T, C> {
             }
 
             // -------------------- "Core" Functionality -------------------- //
-            // TODO: Improve the '?' response based on last-sent stop reason.
-            // this will be particularly relevant when working on non-stop mode.
-            Base::QuestionMark(_) => self.report_reasonable_stop_reason(res, target)?,
+            Base::QuestionMark(_) => {
+                // TODO: Improve the '?' response.
+                // this will be particularly relevant when working on non-stop mode.
+                self.report_reasonable_stop_reason(res, target)?
+            }
             Base::qAttached(cmd) => {
                 let is_attached = match target.support_extended_mode() {
                     // when _not_ running in extended mode, just report that we're attaching to an
@@ -349,62 +360,9 @@ impl<T: Target, C: Connection> GdbStubImpl<T, C> {
                 }
                 HandlerStatus::NeedsOk
             }
-            Base::qC(_) => {
-                res.write_str("QC")?;
-                let pid = self.get_sane_current_pid(target)?;
-
-                match target.base_ops() {
-                    BaseOps::SingleThread(_) => res.write_specific_thread_id(SpecificThreadId {
-                        pid: self
-                            .features
-                            .multiprocess()
-                            .then_some(SpecificIdKind::WithId(pid)),
-                        tid: SpecificIdKind::WithId(SINGLE_THREAD_TID),
-                    })?,
-                    BaseOps::MultiThread(ops) => {
-                        if self.current_mem_tid == SINGLE_THREAD_TID {
-                            let mut err: Result<_, Error<T::Error, C::Error>> = Ok(());
-                            let mut first = true;
-                            ops.list_active_threads(&mut |tid| {
-                                // TODO: replace this with a try block (once stabilized)
-                                let e = (|| {
-                                    if !first {
-                                        return Ok(());
-                                    }
-                                    first = false;
-                                    res.write_specific_thread_id(SpecificThreadId {
-                                        pid: self
-                                            .features
-                                            .multiprocess()
-                                            .then_some(SpecificIdKind::WithId(pid)),
-                                        tid: SpecificIdKind::WithId(tid),
-                                    })?;
-                                    Ok(())
-                                })();
-
-                                if let Err(e) = e {
-                                    err = Err(e)
-                                }
-                            })
-                            .map_err(Error::TargetError)?;
-                            err?
-                        } else {
-                            res.write_specific_thread_id(SpecificThreadId {
-                                pid: self
-                                    .features
-                                    .multiprocess()
-                                    .then_some(SpecificIdKind::WithId(pid)),
-                                tid: SpecificIdKind::WithId(self.current_mem_tid),
-                            })?;
-                        }
-                    }
-                }
-
-                HandlerStatus::Handled
-            }
             Base::qfThreadInfo(_) => {
                 res.write_str("m")?;
-                let pid = self.get_sane_current_pid(target)?;
+                let pid = self.get_current_pid(target)?;
 
                 match target.base_ops() {
                     BaseOps::SingleThread(_) => res.write_specific_thread_id(SpecificThreadId {