@daniel5151 2023-03-25
@xobs after finding some time to play around with vAttach in earnest, I ended up making some changes to the implementation from #129.

Can you please try out this PR and let me know that it works as intended? And of course - feel free to leave any review comments on it.