@xobs 2023-03-26
Happy to see the removal of the comment about it being an untested code path.

Once the Renode changes land, I'll finish up the kernel stuff and then share instructions on how to use gdbstub with the emulator. It should also work with real hardware at that point.