@daniel5151 2023-03-27
Are you sure that error is from `qC` not being implemented, or that the stub has crashed with a `MissingCurrentActivePidImpl` error?

Because the code _should_ throw that error if you try and use vAttach without having implemented the `CurrentActivePid` IDET...