@xobs 2023-03-27
The stub didn't crash, the issue definitely came from within gdb. It was similar to what I experienced when I didn't have `qC` implemented before.