@daniel5151 2023-03-27
Are you certain of that? Because those

```
[remote] getpkt_or_notif_sane_1: Timed out.
[remote] getpkt_or_notif_sane_1: Timed out.
[remote] getpkt_or_notif_sane_1: Timed out.
Ignoring packet error, continuing...
```

messages make me think that the stub died...

Any chance you could try and repro using the in-tree armv4t example? i.e: comment out the `support_current_active_pid()` implementation, and see if it still happens.