@xobs 2023-03-27
I won't be able to test it for another week -- I'm attending Latchup in Santa Barbara this weekend.