@daniel5151 2023-03-28
Sure, no rush.

Also, when you get the chance: can you let me know what GDB client version you're using?