@xobs 2023-04-05
I had a chance to test it, and you're right in that it's caused by the error:

```
(base) user@Ondo:/opt/Xous/gdbstub/examples/armv4t$ cargo run --example armv4t
    Finished dev [unoptimized + debuginfo] target(s) in 0.01s
     Running `/opt/Xous/gdbstub/target/debug/examples/armv4t`
loading section ".text" into memory from [0x55550000..0x55550078]
Setting PC to 0x55550000
Waiting for a GDB connection on "127.0.0.1:9001"...
Debugger connected from 127.0.0.1:52558
GDB queried if it was attached to a process with PID 1
GDB sent a kill request for pid Some(1)
gdbstub encountered a fatal error: GDB client attempted to attach to a new process, but the target has not implemented support for `ExtendedMode::support_current_active_pid`
Program completed. Return value: 0
```

This wasn't caught in my system because unlike `armv4t`, my connection goes over a serial port which does not have any concept of "Remote connection closed". Thus, when it panics like that, it immediately restarts and is in a bad state when gdb sends `$vAttach;2#38` and the stub isn't expecting it.

Ordinarily I'm testing riscv stuff with https://github.com/xpack-dev-tools/riscv-none-elf-gcc-xpack/ but for this I'm using `gdb-multiarch` from Ubuntu 22.04 (version 12.1).