@daniel5151 2023-04-06
Yeah, that lines up with what I expected.

`gdbstub` explicitly works _really hard_ to avoid introducing any panic paths to avoid these sorts of situations where things suddenly die without explanation.

I suggest making sure that ensure `gdbstub` errors are properly logged before panicking.

* * * 

In any case though, with that out of the way - does this code look good to you? Any other issues, or can I merge it in?